Deduplicate SparseIndex
